Add files via upload
Assembly.v3 contains everything from Assembly.v4. It is recommended to download the previous branches to rebuild OpenCALAssembly.v3 file.

Assembly.v3 also contains a new vertical 20100 beam. This design was not deemed perfect, and will be replaced with a NEMA 17 Lead Screw for Z-stage motion. For users who may be fine with the rigidity of the 20100 beam and prefer avoiding Z-stage motion, this may be a good stage to test with. DISCLAIMER: THIS HAS NOT BEEN ASSEMBLED OR TESTED YET BY THE DESIGNERS