Add files via upload
This is a fully parameterized collet for coupling to vials. 

To change the dimensions, Tools --> Equations --> Insert O.D. of vial (or vial cap) in "viam diam" 

The max allowable vial cap O.D is 40mm to allow for coupling with the tool holder. For larger sizes, it may be preferable to increase collet length in  "colletheight"

Our lab is currently experimenting with various AM materials to ensure that the collet design works.