Add files via upload
Minor tolerancing changes included in this assembly, as well as a new collet standoff after experimenting in the lab. 

1. Aluminum/3D printable standoff for Stepper Motor (printing this as part of the frame produced a weak point, so I'd recommend printing this on a Markforged or machining it on a mill (if you have access to one.)

I'll be putting in a branch "Tutorials" on how to adjust various parameters such as screw sizes to mate with the linear actuator stage, adjust collet diameters for different vials, and setting up the hardware. Stay tuned on this front as we begin rolling out further development. 

If you have questions/queries/frustrations, please always feel free to reach out to me at tristanschwab@berkeley.edu. I love working to make this platform easy to use and accessible. Cheers!